Updated Project
Added constraints file with the proper verification with the board schematic and the signals definition, modified the adn2814 signal that is present in the DAPHNE_V2 board and made the correct mapping to the cdr signal present in the DAPHNE_V1 board. Also created top module in the project containing all of these signals. Also added a package file