Finalized DAPHNE V1 Design
The first version of the design for the DAPHNE V1 Board has been finished, including the initial clock and data retrieval for and from the AFE5808As, including the self trigger module, as well as the ethernet part that connects with the MAC and PHY elements for the connection. Several timing issues were also fixed, including CDC, bus skews and more. The design only has a few warning sbut are related to unused signals around the module. There are no critical warnings nor errors and all bus and sinmple path timing constraints are met. The bitstream has been succesfully generated.