Tested the design for the first time
After hours of messing around, the design was finally tested succesfully up to the point where the data is properly alligned by using the automatic allignment method. There are a few things to consider still regarding the ethernet but looks promising since the PHY link is stablished with the router. Must test next week specially since the commands coming from the terminal are not properly working sometimes.