DAPHNE V1 Projects Modifications
Modified the DAPHNE_Full project located at the "Old Versions" folder since its migration messed with the design sources path (Fixed).
Created the daphnev1 top module in the new definitive project, including the addition of the timing endpoint for the main clocks generation, the addition fo the system constraints, and the addition of both AcquisitionManager and Ethernet Modules, still need to add the DAPHNE V1 Board constraints and the Self Trigger Module, which may be bugged because of the new frequencies of operation, as well as the triggering template used (Since the one used in DAPHNE_Full project had an undersampled one).