Finished with the main Design
The whole design was finally finished and is completely linked on its tree of files. there does not seem to be any inconsistency as of now, and both the self trigger and the filter have had its template and its coefficients modified accordingly. The system is ready to be tested, although there could be some tests in simulation for the AXI Stream side of the self trigger