Tested daphne1 Project
The whole project was tested and its bitstream is attached to this commit. The reception part of the system was properly verified and is working as intended, the LEDs are showing that any key that is pressed in the kayboard is stored in the FIFO nad shown in these LEDs. Tghe same logic can be used to any string of data that is sent. But in that case it is picking the first Payload Byte that it receives. Must test now the Tx Section of the Module.