Functional DAPHNE V1 Design
The VHDL Version for DAPHNE V1 is finally working as intended with minor fixes needed, regarding the external trigger generation with the ethernet connection, although it is possible, still is happening one commnd later, so this will be addressed in future versions of the code. Data Alignment is functional, self trigger is also functional, and both Rx and Tx of ethernet are functional.