Added new rounding method for the HIgh Pass FIlter
Added a new rounding method for the HIgh Pass Filter that uses a rounding condition, isntead of always rounding down by using the sifht right operation, if the output of the filter has a fractional part that is bigger or equal to 0.5, it rounds the result to the round down version + 1, this means that we are now rounding to the nearest integer. This still needs to be verified in simulation!