Succesfully generated bitstream
After modifying the neural network file and solved the inferred latch in the filter's outputs, the bitstream was succesfully generated for the DAPHNE V1 board design featuring a neural network based self trigger. Must test with the DAPHNE board, but seems promising. Before any other test, we must verify a new way of implementing both filters in a pipelined structure!