Tested new self trigger conditions
Modified the self trigger algorithm in order to stop triggering with a neural network or cross correlation algorithm, now the system only triggers by manually sending the letter "t" through ethernet. The system works as intended but the self trigger features have not been able to be tsted since the noise levels in the Board are way too high so every single window is considered an event window, must be fixed later!