Succesfully tested both algorithms
Both self trigger algorithms (Cross-correlation and Neural Networks) have been succesfully tested in the DAPHNE V1 Board. The bitstream is included in this commit, which contains the Neural Network as the self trigger, with a Threshold level of 1. With this level, the network is capable of detecting 10000 out of 10000 events generated with the Analog Discovery 2 device, meaning a 100% success rate, just like with the cross correlation algorithm, there are indeed some cases when the system triggers with random noise, but by inspecting the noise it looks almost like an event.