Added real tests folder
Added a folder that includes all the tests that were done for the two different self trigger algorithms that EIA designed for the PDS Consortium and that were implemented in the DAPHNE V1 Board. They can be considered as successful, and the packets that were used are all captured in the Wireshark .pcapng files. Images and their respective MATLAB figures were generated for some of the pakcets that were randomly selected. Also added a .txt file that includes the configuration of the AFE, as well as the configuration of the event signal in the WaveForms software for the Analog Disovery 2 Device used to generate the trigger signal/event signal