Added new ethernet commands
Added new ethernet commands to "manually" control the allignment process of the FPGA. As of right now, the bitstream of this last modification is missing, but the allignment module was set to operate with the manual allignment, for this it needs very specifica training patterns that the AFE can output by using the UART communication. This patterns can be output in the deskew and then custom order so that the data is properly alligned. The finite state machines that hold this task will infinitely be doing their process until this patterns are seen at the output of the AFE. After succesfully testing this feature, the signals created to generate this commands from ethernet will be connected to the module and another test will be performed.